+++ litestar/contrib/repository/__init__.py
@@ -1,7 +1,9 @@
 from __future__ import annotations
 
-from .abc import AbstractAsyncRepository, AbstractSyncRepository, FilterTypes
+from .async_base import AbstractAsyncRepository
 from .exceptions import ConflictError, NotFoundError, RepositoryError
+from .filters import FilterTypes
+from .sync_base import AbstractSyncRepository
 
 __all__ = (
     "AbstractAsyncRepository",
