+++ litestar/contrib/repository/testing/generic_mock_repository.py
@@ -8,14 +8,13 @@ from datetime import datetime, timezone, tzinfo
 from typing import TYPE_CHECKING, Generic, Protocol, TypeVar
 from uuid import uuid4
 
-from litestar.contrib.repository.abc import AbstractAsyncRepository, AbstractSyncRepository
+from litestar.contrib.repository import AbstractAsyncRepository, AbstractSyncRepository, FilterTypes
 from litestar.contrib.repository.exceptions import ConflictError, RepositoryError
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Hashable, Iterable, MutableMapping
     from typing import Any
 
-    from litestar.contrib.repository import FilterTypes
 
 ModelT = TypeVar("ModelT", bound="HasID")
 AsyncMockRepoT = TypeVar("AsyncMockRepoT", bound="GenericAsyncMockRepository")
